feat: 응

+++ GustomateApp/preference/schema.py
@@ -2,8 +2,8 @@ from pydantic import BaseModel
 from typing import List
 
 class UserPreferenceBase(BaseModel):
-    spiciness_preference: int
-    cooking_skill: int
+    spiciness_preference: float
+    cooking_skill: float
     is_on_diet: bool
     has_allergies: bool  # 알러지 여부
     allergies: List[str] = []  # 알러지 목록
